requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
protobuf==3.13.0
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
